requirements.txt
streamlit==1.32.0
pandas
feedparser
google-genai
python-dotenv